Update for 3.12.2021 - Uploading HEX and JS Files
This is the first update in this repository, which adds the BBC micro:bit compatible file (.hex) and the JavaScript file for the source code (.js).
----------------------
- How do I edit these?

 If you want to edit the source code through an editor that is not Microsoft Makecode (ex. Microsoft Visual Studio), please use the .js file, as it will automatically fill all the variables and the colours of the code with it. If you want to edit the files in Makecode, please copy the code from the separate file into the editor and change the mode to the corresponding mode.
To upload the .hex file to your micro:bit board (the code supports gen. 1 and gen. 2 versions of the board), plug the board in your computer and move the file to the icon of the micro:bit. It will upload automatically. Wait until prompted.

Thank you for downloading!

Copyright alokinBlogs 2020-2022, ARR.